refactor(iiifparser): fold the parse_iiif_uri classifier out of src/handlers
`src/handlers/` held one file — the C++ `parse_iiif_uri` request classifier —
under a directory name (DUNE-014) that promised more than a single URI parser.
The Rust shell parses IIIF URIs in production (`server-rs/src/iiif.rs`, a port
of this classifier), so the C++ code is now only the reference its unit,
approval, and corpus tests check against.

git-mv the two files into the carved `//src/iiifparser` package as a new
`testonly` `:iiif_handler` target, beside the IIIF component parsers it
classifies into. Because it is `testonly` and out of the `//src:sipi_lib` glob,
this also drops the dead classifier from the production binary; the three test
targets depend on the narrow target directly. Retire the `handlers` module scope
(folded into `iiifparser`) in CONVENTIONS.md and the commit-scope vocabulary.

The `//fuzz/handlers` half of DUNE-014 was already handled in Phase 7 (the fuzz
target + tree were deleted with the oracle); nothing recreated them.
