+++ test/approval/BUILD.bazel
@@ -48,6 +48,10 @@ cc_test(
     local_defines = ["APPROVAL_TESTS_HIDE_DEPRECATED_CODE=1"],
     deps = [
         "//src:sipi_lib",
+        # The `parse_iiif_uri` classifier `iiif_parse_approval_test.cpp` checks;
+        # testonly, no longer in `//src:sipi_lib` (folded into //src/iiifparser,
+        # DUNE-014).
+        "//src/iiifparser:iiif_handler",
         "//test:test_paths",
         "@approvaltests_cpp//:approval_tests",
         "@googletest//:gtest_main",
