+++ test/approval/iiif_parse_approval_test.cpp
@@ -1,7 +1,7 @@
 #include "ApprovalTests.hpp"
 #include "gtest/gtest.h"
 
-#include "handlers/iiif_handler.h"
+#include "iiifparser/iiif_handler.h"
 
 #include <sstream>
