+++ test/unit/handlers/iiif_handler_test.cpp
@@ -11,7 +11,7 @@
 #include <string>
 #include <thread>
 
-#include "handlers/iiif_handler.h"
+#include "iiifparser/iiif_handler.h"
 
 using namespace handlers::iiif_handler;
