+++ test/unit/handlers/parse_iiif_uri_corpus_test.cpp
@@ -17,7 +17,7 @@
 #include <string>
 #include <thread>
 
-#include "handlers/iiif_handler.h"
+#include "iiifparser/iiif_handler.h"
 
 #ifndef SIPI_FUZZ_CORPUS_DIR
 #error "SIPI_FUZZ_CORPUS_DIR must point at test/unit/handlers/corpus/"
